Extract topics widget rendering into TopicsToDiscoverNextDelegate
